Rpc login now throws custom checked exception. Little refactoring.
