+++ admin-service-api/src/main/java/hu/schonherz/project/admin/service/api/rpc/FailedRpcLoginAttemptException.java
@@ -1,0 +1,13 @@
+package hu.schonherz.project.admin.service.api.rpc;
+
+public class FailedRpcLoginAttemptException extends Exception {
+
+    public FailedRpcLoginAttemptException(String message) {
+        super(message);
+    }
+
+    public FailedRpcLoginAttemptException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
