+++ admin-service-api/src/main/java/hu/schonherz/project/admin/service/api/rpc/RpcLoginServiceRemote.java
@@ -2,6 +2,6 @@ package hu.schonherz.project.admin.service.api.rpc;
 
 public interface RpcLoginServiceRemote {
 
-    String rpcLogin(String username, String plainTextPassword);
+    String rpcLogin(String username, String plainTextPassword) throws FailedRpcLoginAttemptException;
 
 }
