+++ admin-service-api/src/main/java/hu/schonherz/project/admin/service/api/service/UserServiceRemote.java
@@ -1,6 +1,5 @@
 package hu.schonherz.project.admin.service.api.service;
 
-import hu.schonherz.project.admin.service.api.service.exception.InvalidUserDataException;
 import hu.schonherz.project.admin.service.api.vo.UserVo;
 
 public interface UserServiceRemote extends UserService {
