+++ admin-service/src/main/java/hu/schonherz/project/admin/service/impl/user/RpcLoginServiceBean.java
@@ -1,6 +1,7 @@
 package hu.schonherz.project.admin.service.impl.user;
 
 import hu.schonherz.project.admin.service.api.encrypter.Encrypter;
+import hu.schonherz.project.admin.service.api.rpc.FailedRpcLoginAttemptException;
 import hu.schonherz.project.admin.service.api.rpc.RpcLoginServiceRemote;
 import hu.schonherz.project.admin.service.api.service.UserServiceLocal;
 import hu.schonherz.project.admin.service.api.vo.UserVo;
@@ -9,6 +10,7 @@ import javax.ejb.Remote;
 import javax.ejb.Stateless;
 import javax.jws.WebMethod;
 import javax.jws.WebService;
+import lombok.NonNull;
 
 @Stateless(mappedName = "RpcLoginService")
 @Remote(UserServiceLocal.class)
@@ -20,21 +22,25 @@ public class RpcLoginServiceBean implements RpcLoginServiceRemote {
 
     @Override
     @WebMethod(operationName = "login")
-    public String rpcLogin(String username, String plainTextPassword) {
+    public String rpcLogin(@NonNull String username, @NonNull String plainTextPassword) throws FailedRpcLoginAttemptException {
+        if (username.isEmpty() || plainTextPassword.isEmpty()) {
+            throw new IllegalArgumentException("Username and password must not be empty string!");
+        }
+
         UserVo user = userService.findByUsername(username);
         if (user == null) {
-            throw new IllegalArgumentException("The user " + username + " does not exist!");
+            throw new FailedRpcLoginAttemptException("The user " + username + " does not exist!");
         }
 
         String encryptedGivenPassword = Encrypter.encrypt(plainTextPassword);
         String encryptedUserPassword = user.getPassword();
 
         if (!Encrypter.match(encryptedUserPassword, encryptedGivenPassword)) {
-            throw new IllegalArgumentException("Invalid password!");
+            throw new FailedRpcLoginAttemptException("Invalid password!");
         }
 
         if (!user.isActive()) {
-            throw new IllegalArgumentException("User " + username + " is inactive!");
+            throw new FailedRpcLoginAttemptException("User " + username + " is inactive!");
         }
 
         return user.getUserRole();
